shared lives in core
