--- core/examples/basic/.gitignore
@@ -1,1 +1,0 @@
-/target
