--- core/examples/basic/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "basic"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-pgmq = { path = "../../"}
-serde = "1.0.152"
-serde_json = "1.0.93"
-tokio = "1.25.0"
